Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -2,8 +2,8 @@
 
 
 int main() {
-    int x;
-    scanf("%d",&x);
+    float x;
+    scanf("%f",&x);
     float area;
     area=x*x*3.14;
     printf("Area: %.2f",area);
